Rename user repository
