+++ src/main/kotlin/com/tradingplatform/data/UserRepository.kt
@@ -1,0 +1,15 @@
+package com.tradingplatform.data
+
+import com.tradingplatform.model.User
+
+object UserRepository {
+    var users = HashMap<String, User>()
+
+    fun addUser(user: User) {
+        users[user.userName] = user
+    }
+
+    fun getUser(userName: String): User? {
+        return users[userName]
+    }
+}
