+++ src/test/kotlin/com/tradingplatform/InventoryTest.kt
@@ -1,7 +1,7 @@
 package com.tradingplatform
 
 import com.tradingplatform.controller.InventoryController
-import com.tradingplatform.data.UserRepo
+import com.tradingplatform.data.UserRepository
 import com.tradingplatform.model.User
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
@@ -12,7 +12,7 @@ class InventoryTest {
     fun `check whether Performance esops added to inventory`() {
 
         val user = User("", "", "", "tt@gmail.com", "atul_1")
-        UserRepo.addUser(user)
+        UserRepository.addUser(user)
         val objectOfInventoryController = InventoryController()
 
 
@@ -27,7 +27,7 @@ class InventoryTest {
     fun `check whether Normal esops added to inventory`() {
 
         val user = User("", "", "", "tt@gmail.com", "atul_1")
-        UserRepo.addUser(user)
+        UserRepository.addUser(user)
         val objectOfInventoryController = InventoryController()
