+++ src/test/kotlin/com/tradingplatform/controller/WalletControllerTest.kt
@@ -1,6 +1,6 @@
 package com.tradingplatform.controller
 
-import com.tradingplatform.data.UserRepo
+import com.tradingplatform.data.UserRepository
 import com.tradingplatform.model.*
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest
 import io.restassured.specification.RequestSpecification
@@ -13,13 +13,13 @@ class WalletControllerTest {
 
     @BeforeEach
     fun `Remove all the Users and Orders`() {
-        UserRepo.users.clear()
+        UserRepository.users.clear()
     }
     @Test
     fun `valid amount entered on post request`(spec: RequestSpecification) {
 
         val user = User("", "", "", "tat@gmail.com", "sahaj")
-        UserRepo.addUser(user)
+        UserRepository.addUser(user)
         spec.`when`()
             .header("Content-Type", "application/json")
             .body("{\"amount\": 10}")
@@ -33,7 +33,7 @@ class WalletControllerTest {
     @Test
     fun `amount entered is negative`(spec: RequestSpecification) {
         val user = User("", "", "", "tat@gmail.com", "pkcs")
-        UserRepo.addUser(user)
+        UserRepository.addUser(user)
         spec.`when`()
             .header("Content-Type", "application/json")
             .body("{\"amount\": -10}")
@@ -48,7 +48,7 @@ class WalletControllerTest {
     @Test
     fun `amount entered exceeds maxLimit`(spec: RequestSpecification) {
         val user = User("", "", "", "tat@gmail.com", "pkcs")
-        UserRepo.addUser(user)
+        UserRepository.addUser(user)
         spec.`when`()
             .header("Content-Type", "application/json")
             .body("{\"amount\": 10000001}")
@@ -64,7 +64,7 @@ class WalletControllerTest {
     fun `amount entered is a string`(spec: RequestSpecification) {
 
         val user = User("", "", "", "tat@gmail.com", "pkcs")
-        UserRepo.addUser(user)
+        UserRepository.addUser(user)
         spec.`when`()
             .header("Content-Type", "application/json")
             .body("{\"amount\": \"abc\"}")
@@ -78,7 +78,7 @@ class WalletControllerTest {
     @Test
     fun `amount field not entered`(spec: RequestSpecification) {
         val user = User("", "", "", "tat@gmail.com", "pkcs")
-        UserRepo.addUser(user)
+        UserRepository.addUser(user)
         spec.`when`()
             .header("Content-Type", "application/json")
             .body("{}")
